+++ grid2op/Backend/pandaPowerBackend.py
@@ -327,7 +327,8 @@ class PandaPowerBackend(Backend):
             warnings.simplefilter("ignore", FutureWarning)
             self._grid = copy.deepcopy(self.__pp_backend_initial_grid)
         self._reset_all_nan()
-        self._topo_vect[:] = self._get_topo_vect()
+        self._get_line_status()
+        self._get_topo_vect()
         self.comp_time = 0.0
 
     def load_grid(self,
@@ -717,6 +718,8 @@ class PandaPowerBackend(Backend):
             )
         
         self._compute_pos_big_topo()
+        
+        self._topo_vect = np.full(self.dim_topo, fill_value=-1, dtype=dt_int)
 
         # utilities for imeplementing apply_action
         self._corresp_name_fun = {}
@@ -768,7 +771,6 @@ class PandaPowerBackend(Backend):
         self.q_ex = np.full(self.n_line, dtype=dt_float, fill_value=np.NaN)
         self.v_ex = np.full(self.n_line, dtype=dt_float, fill_value=np.NaN)
         self.a_ex = np.full(self.n_line, dtype=dt_float, fill_value=np.NaN)
-        self.line_status = np.full(self.n_line, dtype=dt_bool, fill_value=np.NaN)
         self.load_p = np.full(self.n_load, dtype=dt_float, fill_value=np.NaN)
         self.load_q = np.full(self.n_load, dtype=dt_float, fill_value=np.NaN)
         self.load_v = np.full(self.n_load, dtype=dt_float, fill_value=np.NaN)
@@ -779,6 +781,9 @@ class PandaPowerBackend(Backend):
         self.storage_q = np.full(self.n_storage, dtype=dt_float, fill_value=np.NaN)
         self.storage_v = np.full(self.n_storage, dtype=dt_float, fill_value=np.NaN)
         self._nb_bus_before = None
+        
+        self.line_status = np.full(self.n_line, dtype=dt_bool, fill_value=np.NaN)
+        self.line_status.flags.writeable = False
 
         # store the topoid -> objid
         self._init_topoid_objid()
@@ -805,7 +810,7 @@ class PandaPowerBackend(Backend):
         self.gen_theta = np.full(self.n_gen, fill_value=np.NaN, dtype=dt_float)
         self.storage_theta = np.full(self.n_storage, fill_value=np.NaN, dtype=dt_float)
 
-        self._topo_vect = self._get_topo_vect()
+        self._get_topo_vect()
         self.tol = 1e-5  # this is NOT the pandapower tolerance !!!! this is used to check if a storage unit
         # produce / absorbs anything
 
@@ -824,7 +829,10 @@ class PandaPowerBackend(Backend):
         self.storage_p = np.full(cls.n_storage, dtype=dt_float, fill_value=np.NaN)
         self.storage_q = np.full(cls.n_storage, dtype=dt_float, fill_value=np.NaN)
         self.storage_v = np.full(cls.n_storage, dtype=dt_float, fill_value=np.NaN)
-        self._topo_vect = self._get_topo_vect()
+        self._topo_vect.flags.writeable = True
+        self._topo_vect.resize(cls.dim_topo)
+        self._topo_vect.flags.writeable = False
+        self._get_topo_vect()
 
     def _convert_id_topo(self, id_big_topo):
         """
@@ -894,7 +902,6 @@ class PandaPowerBackend(Backend):
             tmp_stor_p = self._grid.storage["p_mw"]
             if (storage.changed).any():
                 tmp_stor_p.iloc[storage.changed] = storage.values[storage.changed]
-
             # topology of the storage
             stor_bus = backendAction.get_storages_bus()
             new_bus_num = dt_int(1) * self._grid.storage["bus"].values
@@ -906,11 +913,7 @@ class PandaPowerBackend(Backend):
             self._grid.storage.loc[stor_bus.changed & deactivated, "in_service"] = False
             self._grid.storage.loc[stor_bus.changed & ~deactivated, "in_service"] = True
             self._grid.storage["bus"] = new_bus_num
-            self._topo_vect[cls.storage_pos_topo_vect[stor_bus.changed]] = new_bus_id
-            self._topo_vect[
-                cls.storage_pos_topo_vect[deact_and_changed]
-            ] = -1
-
+        
         if type(backendAction).shunts_data_available:
             shunt_p, shunt_q, shunt_bus = shunts__
 
@@ -936,6 +939,8 @@ class PandaPowerBackend(Backend):
             if type_obj is not None:
                 # storage unit are handled elsewhere
                 self._type_to_bus_set[type_obj](new_bus, id_el_backend, id_topo)
+        
+        self._topo_vect.flags.writeable = False
 
     def _apply_load_bus(self, new_bus, id_el_backend, id_topo):
         new_bus_backend = type(self).local_bus_to_global_int(
@@ -1083,7 +1088,14 @@ class PandaPowerBackend(Backend):
         in case of "do nothing" action applied.
         """
         try:
+            # as pandapower does not modify the topology or the status of 
+            # powerline, then we can compute the topology (and the line status)
+            # at the beginning
+            # This is also interesting in case of divergence :-)
+            self._get_line_status()
+            self._get_topo_vect()
             self._aux_runpf_pp(is_dc)
+            
             cls = type(self)     
             # if a connected bus has a no voltage, it's a divergence (grid was not connected)
             if self._grid.res_bus.loc[self._grid.bus["in_service"]]["va_degree"].isnull().any():
@@ -1128,8 +1140,7 @@ class PandaPowerBackend(Backend):
                             ):
                                 self.load_v[l_id] = self.prod_v[g_id]
                                 break
-                            
-            self.line_status[:] = self._get_line_status()
+            
             # I retrieve the data once for the flows, so has to not re read multiple dataFrame
             self.p_or[:] = self._aux_get_line_info("p_from_mw", "p_hv_mw")
             self.q_or[:] = self._aux_get_line_info("q_from_mvar", "q_hv_mvar")
@@ -1174,17 +1185,12 @@ class PandaPowerBackend(Backend):
                 self.storage_v[:],
                 self.storage_theta[:],
             ) = self._storages_info()
+            
             deact_storage = ~np.isfinite(self.storage_v)
-            if (np.abs(self.storage_p[deact_storage]) > self.tol).any():
-                raise pp.powerflow.LoadflowNotConverged(
-                    "Isolated storage set to absorb / produce something"
-                )
             self.storage_p[deact_storage] = 0.0
             self.storage_q[deact_storage] = 0.0
             self.storage_v[deact_storage] = 0.0
             self._grid.storage["in_service"].values[deact_storage] = False
-
-            self._topo_vect[:] = self._get_topo_vect()
             if not self._grid.converged:
                 raise pp.powerflow.LoadflowNotConverged("Divergence without specific reason (self._grid.converged is False)")
             self.div_exception = None
@@ -1222,7 +1228,12 @@ class PandaPowerBackend(Backend):
         self.load_theta[:] = np.NaN
         self.gen_theta[:] = np.NaN
         self.storage_theta[:] = np.NaN
-
+        self._topo_vect.flags.writeable = True
+        self._topo_vect[:] = -1
+        self._topo_vect.flags.writeable = False
+        self.line_status.flags.writeable = True
+        self.line_status[:] = False
+        self.line_status.flags.writeable = False
     def copy(self) -> "PandaPowerBackend":
         """
         INTERNAL
@@ -1324,7 +1335,7 @@ class PandaPowerBackend(Backend):
         res._in_service_trafo_col_id = self._in_service_trafo_col_id
         
         res._missing_two_busbars_support_info = self._missing_two_busbars_support_info
-        res._missing_detachment_support = self._missing_detachment_support
+        res._missing_detachment_support_info = self._missing_detachment_support_info
         res.div_exception = self.div_exception
         return res
 
@@ -1368,12 +1379,15 @@ class PandaPowerBackend(Backend):
         return self.line_status
 
     def _get_line_status(self):
-        return np.concatenate(
+        self.line_status.flags.writeable = True
+        self.line_status[:] = np.concatenate(
             (
                 self._grid.line["in_service"].values,
                 self._grid.trafo["in_service"].values,
             )
         ).astype(dt_bool)
+        self.line_status.flags.writeable = False
+        return self.line_status
 
     def get_line_flow(self) -> np.ndarray:
         return self.a_or
@@ -1383,45 +1397,62 @@ class PandaPowerBackend(Backend):
             self._grid.line.iloc[id_, self._in_service_line_col_id] = False
         else:
             self._grid.trafo.iloc[id_ - self._number_true_line, self._in_service_trafo_col_id]  = False
+        self._topo_vect.flags.writeable = True
         self._topo_vect[self.line_or_pos_topo_vect[id_]] = -1
         self._topo_vect[self.line_ex_pos_topo_vect[id_]] = -1
+        self._topo_vect.flags.writeable = False
+        self.line_status.flags.writeable = True
         self.line_status[id_] = False
+        self.line_status.flags.writeable = False
 
     def _reconnect_line(self, id_):
         if id_ < self._number_true_line:
             self._grid.line.iloc[id_, self._in_service_line_col_id] = True
         else:
             self._grid.trafo.iloc[id_ - self._number_true_line, self._in_service_trafo_col_id] = True
+        self.line_status.flags.writeable = True
         self.line_status[id_] = True
+        self.line_status.flags.writeable = False
 
     def get_topo_vect(self) -> np.ndarray:
         return self._topo_vect
 
     def _get_topo_vect(self):
-        cls = type(self)
-        res = np.full(cls.dim_topo, fill_value=np.iinfo(dt_int).max, dtype=dt_int)
+        """
+        .. danger:: 
+            you should have called `self._get_line_status` before otherwise it might
+            not behave correctly !
 
+        Returns
+        -------
+        _type_
+            _description_
+        """
+        cls = type(self)
+        
         # lines / trafo
         line_status = self.get_line_status()
+        self._topo_vect.flags.writeable = True
         glob_bus_or = np.concatenate((self._grid.line["from_bus"].values, self._grid.trafo["hv_bus"].values))
-        res[cls.line_or_pos_topo_vect] = cls.global_bus_to_local(glob_bus_or, cls.line_or_to_subid)
-        res[cls.line_or_pos_topo_vect[~line_status]] = -1
+        self._topo_vect[cls.line_or_pos_topo_vect] = cls.global_bus_to_local(glob_bus_or, cls.line_or_to_subid)
+        self._topo_vect[cls.line_or_pos_topo_vect[~line_status]] = -1
         glob_bus_ex = np.concatenate((self._grid.line["to_bus"].values, self._grid.trafo["lv_bus"].values))
-        res[cls.line_ex_pos_topo_vect] = cls.global_bus_to_local(glob_bus_ex, cls.line_ex_to_subid)
-        res[cls.line_ex_pos_topo_vect[~line_status]] = -1
+        self._topo_vect[cls.line_ex_pos_topo_vect] = cls.global_bus_to_local(glob_bus_ex, cls.line_ex_to_subid)
+        self._topo_vect[cls.line_ex_pos_topo_vect[~line_status]] = -1
         # load, gen
         load_status = self._grid.load["in_service"].values
-        res[cls.load_pos_topo_vect] = cls.global_bus_to_local(self._grid.load["bus"].values, cls.load_to_subid)
-        res[cls.load_pos_topo_vect[~load_status]] = -1
+        self._topo_vect[cls.load_pos_topo_vect] = cls.global_bus_to_local(self._grid.load["bus"].values, cls.load_to_subid)
+        self._topo_vect[cls.load_pos_topo_vect[~load_status]] = -1
         gen_status = self._grid.gen["in_service"].values
-        res[cls.gen_pos_topo_vect] = cls.global_bus_to_local(self._grid.gen["bus"].values, cls.gen_to_subid)
-        res[cls.gen_pos_topo_vect[~gen_status]] = -1
+        self._topo_vect[cls.gen_pos_topo_vect] = cls.global_bus_to_local(self._grid.gen["bus"].values, cls.gen_to_subid)
+        self._topo_vect[cls.gen_pos_topo_vect[~gen_status]] = -1
         # storage
         if cls.n_storage:
             storage_status = self._grid.storage["in_service"].values
-            res[cls.storage_pos_topo_vect] = cls.global_bus_to_local(self._grid.storage["bus"].values, cls.storage_to_subid)
-            res[cls.storage_pos_topo_vect[~storage_status]] = -1
-        return res
+            self._topo_vect[cls.storage_pos_topo_vect] = cls.global_bus_to_local(self._grid.storage["bus"].values, cls.storage_to_subid)
+            self._topo_vect[cls.storage_pos_topo_vect[~storage_status]] = -1
+        self._topo_vect.flags.writeable = False
+        return self._topo_vect
 
     def _gens_info(self):
         prod_p = self.cst_1 * self._grid.res_gen["p_mw"].values.astype(dt_float)
@@ -1520,8 +1551,10 @@ class PandaPowerBackend(Backend):
         if self.n_storage:
             # this is because we support "backward comaptibility" feature. So the storage can be
             # deactivated from the Environment...
-            p_storage = self._grid.res_storage["p_mw"].values.astype(dt_float)
-            q_storage = self._grid.res_storage["q_mvar"].values.astype(dt_float)
+            # p_storage = self._grid.res_storage["p_mw"].values.astype(dt_float)
+            # q_storage = self._grid.res_storage["q_mvar"].values.astype(dt_float)
+            p_storage = self._grid.storage["p_mw"].values.astype(dt_float)
+            q_storage = self._grid.storage["q_mvar"].values.astype(dt_float)
             v_storage = (
                 self._grid.res_bus.loc[self._grid.storage["bus"].values][
                     "vm_pu"